+++ app/Main.hs
@@ -3,5 +3,39 @@ module Main
     )
 where
 
+import Control.Monad (when)
+import Control.Concurrent.Async (async, waitAny)
+import Data.Bifunctor (first)
+import Data.Either (rights)
+import DBus.Client (connectSession, requestName, RequestNameReply(..))
+import System.Directory (getXdgDirectory, XdgDirectory(..))
+import System.Exit (exitFailure)
+
+import Config (loadConfig, Config(..), ConfigError(..))
+import Modules.Weather (weatherStringsSvc)
+
 main :: IO ()
-main = putStrLn "hehe"
+main = do
+    -- Read config file
+    path   <- getXdgDirectory XdgConfig "ntfd"
+    cfgRes <- loadConfig $ path <> "/config.toml"
+    config <- case cfgRes of
+        Left e -> do
+            putStrLn $ "Failed to read config file: " <> show e
+            exitFailure
+        Right c -> pure c
+
+    -- Connect to DBus
+    client        <- connectSession
+    requestResult <- requestName client "io.ntfd" []
+    when (requestResult /= NamePrimaryOwner) $ do
+        putStrLn "Another service is connected to DBus as \"io.ntfd\""
+        exitFailure
+
+    -- Prepare services
+    let weatherSvc = first WeatherConfigErr $ weatherStringsSvc client <$> weatherCfg config
+
+    -- Spawn threads and wait forever, or until one of them fails
+    services <- mapM async $ rights [weatherSvc]
+    _        <- waitAny services
+    exitFailure