+++ src/Lib.hs
@@ -1,6 +1,0 @@
-module Lib
-    ( module Config
-    )
-where
-
-import Config