+++ test/Spec/Clients/OpenWeatherMap.hs
@@ -4,6 +4,7 @@ import Control.Concurrent.Async (mapConcurrently)
 import Data.ByteString.Char8 (pack)
 import Data.Either (isRight)
 import Data.Maybe (fromJust)
+import Data.Time.Clock (secondsToNominalDiffTime)
 import System.Environment (lookupEnv)
 import Test.Hspec
 
@@ -14,10 +15,12 @@ defaultCfg :: IO WeatherConfig
 defaultCfg = do
     apiKey <- lookupEnv "OWM_API_KEY"
     pure WeatherConfig
-        { weatherEnabled  = True
-        , weatherApiKey   = fromJust $ pack <$> apiKey
-        , weatherCityId   = "6077243"
-        , weatherTemplate =
+        { weatherEnabled   = True
+        , weatherApiKey    = fromJust $ pack <$> apiKey
+        , weatherCityId    = "6077243"
+        , weatherNotifBody = "hullo"
+        , weatherSyncFreq  = secondsToNominalDiffTime 1800
+        , weatherTemplate  =
             "{{ temp_icon }} {{ temp_celcius }}°C {{ trend }} {{ forecast_icon }} {{ forecast_celcius }}°C"
         }