+++ ChangeLog.md
@@ -1,3 +1,5 @@
 # Changelog for ntfd
 
 ## Unreleased changes
+- Weather module
+- AUR package